enrichment/requirements.txt
python-dotenv
quixstreams==3.16.0
